devop: move how it works layout under core
