+++ src/core/viewDefault/components/howItWorksLayout/HowItWorksDappsCenter.vue
