+++ src/core/viewDefault/components/howItWorksLayout/HowItWorksMore.vue
