+++ src/core/viewDefault/components/howItWorksLayout/HowItWorksSend.vue
