+++ src/core/viewDefault/components/howItWorksLayout/HowItWorksSwap.vue
