+++ src/core/viewDefault/components/howItWorksLayout/HowItWorksTokens.vue
